Extract validator image component
